+++ harmony/RNPing/build-profile.json5
@@ -1,0 +1,16 @@
+{
+  "apiType": "stageMode",
+//  "buildOption": {
+//    "externalNativeOptions": {
+//      "path": "./src/main/cpp/CMakeLists.txt",
+//      "arguments": "",
+//      "cppFlags": ""
+//    },
+//  },
+  "targets": [
+    {
+      "name": "default",
+      "runtimeOS": "HarmonyOS"
+    }
+  ]
+}
