+++ harmony/RNPing/oh-package.json5
@@ -1,0 +1,15 @@
+{
+  "license": "Apache-2.0",
+  "devDependencies": {
+
+  },
+  "author": "",
+  "name": "@react-native-ohos/react-native-ping",
+  "description": "Please describe the basic information.",
+  "main": "index.ets",
+  "version": "1.2.9-rc.1",
+  "dependencies": {
+    "@rnoh/react-native-openharmony":"file:../react_native_openharmony"
+  },
+  "types": 'har'
+}
