+++ harmony/RNPing/src/main/resources/zh_CN/element/string.json
@@ -1,0 +1,8 @@
+{
+  "string": [
+    {
+      "name": "page_show",
+      "value": "page from npm package"
+    }
+  ]
+}
