--- harmony/ping/OAT.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<configuration>
-	<oatconfig>
-	<licensefile>LICENSE</licensefile>
-		<filefilterlist>
-			<filefilter name="copyrightPolicyFilter" desc="Filters for compatibility，license header policies">
-				<filteritem type="filename" name="hvigorfile.ts" desc="hvigor构建脚本，DevEco Studio自动生成，不需要添加版权头"/>
-				<filteritem type="filename" name="*.json5" desc="hvigor工程配置文件，DevEco Studio自动生成，不需要添加版权头"/>
-				<filteritem type="filename" name="*.proto" desc="资源文件，不需要添加版权头"/>
-				<filteritem type="filename" name="*.json" desc="资源文件，不需要添加版权头"/>
-				<filteritem type="filepath" name="hvigorw" desc="工程模板，不修改版权头，以防有修改版权风险"/>
-				<filteritem type="filepath" name="hvigorw.bat" desc="工程模板，不修改版权头，以防有修改版权风险"/>
-				<filteritem type="filepath" name="hvigor/hvigor-wrapper.js" desc="工程模板，不修改版权头，以防有修改版权风险"/>
-                <filteritem type="filename" name="LICENSE" desc="工程文件，不修改版权头"/>
-			</filefilter>
-			<filefilter name="defaultPolicyFilter" desc="Filters for compatibility，license header policies">
-				<filteritem type="filename" name="hvigorfile.ts" desc="hvigor构建脚本，DevEco Studio自动生成，不需要添加许可证头"/>
-				<filteritem type="filename" name="*.json5" desc="hvigor工程配置文件，DevEco Studio自动生成，不需要添加许可证头"/>
-				<filteritem type="filename" name="LICENSE" desc="原三方库证书文件无需更改，因此添加过滤"/>
-				<filteritem type="filename" name="*.proto" desc="资源文件，不需要添加许可证头"/>
-				<filteritem type="filename" name="*.json" desc="资源文件，不需要添加许可证头"/>
-				<filteritem type="filepath" name="hvigorw" desc="工程模板，不修改版权头，以防有修改版权风险"/>
-				<filteritem type="filepath" name="hvigorw.bat" desc="工程模板，不修改版权头，以防有修改版权风险"/>
-				<filteritem type="filepath" name="hvigor/hvigor-wrapper.js" desc="工程模板，不修改版权头，以防有修改版权风险"/>
-			</filefilter>
-			<filefilter name="binaryFileTypePolicyFilter" desc="Filters for resources files policies">
-				<filteritem type="filename" name="icon.png" desc="应用图标"/>
-				<filteritem type="filename" name="app_icon.png" desc="应用图标"/>
-				<filteritem type="filename" name="warn.png" desc="页面展示图标"/>
-			</filefilter>
-		</filefilterlist>
-        <policylist>
-            <policy name="projectPolicy" desc="">
-                <policyitem type="license" name="MIT" path="*.*" desc="license under the MIT"/>
-            </policy>
-        </policylist>
-	</oatconfig>
-</configuration>
