--- harmony/ping/build-profile.json5
@@ -1,16 +1,0 @@
-{
-  "apiType": "stageMode",
-//  "buildOption": {
-//    "externalNativeOptions": {
-//      "path": "./src/main/cpp/CMakeLists.txt",
-//      "arguments": "",
-//      "cppFlags": ""
-//    },
-//  },
-  "targets": [
-    {
-      "name": "default",
-      "runtimeOS": "HarmonyOS"
-    }
-  ]
-}
