--- harmony/ping/hvigorfile.ts
@@ -1,2 +1,0 @@
-// Script for compiling build behavior. It is built in the build plug-in and cannot be modified currently.
-export { harTasks } from '@ohos/hvigor-ohos-plugin';
