--- harmony/ping/index.ets
@@ -1,8 +1,0 @@
-// Copyright (c) 2024 Huawei Device Co., Ltd. All rights reserved
-// Use of this source code is governed by a Apache-2.0 license that can be
-// found in the LICENSE file.
-
-
-import { RNReactNativePingPackage } from './src/main/ets/RNReactNativePingPackage'
-export * from "./ts"
-export default RNReactNativePingPackage
