--- harmony/ping/src/main/ets/Logger.ts
@@ -1,63 +1,0 @@
-/**
- * MIT License
- *
- * Copyright (C) 2023 Huawei Device Co., Ltd.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-import hilog from '@ohos.hilog';
-
-class Logger {
-  private domain: number;
-  private prefix: string;
-  private format: string = '%{public}s, %{public}s';
-  private isDebug: boolean;
-
-  /** constructor.
-   *
-   * @param Prefix Identifies the log tag.
-   * @param domain Domain Indicates the service domain, which is a hexadecimal integer ranging from 0x0 to 0xFFFFF.
-   */
-  constructor(prefix: string = 'MyApp', domain: number = 0xFF00, isDebug = false) {
-    this.prefix = prefix;
-    this.domain = domain;
-    this.isDebug = isDebug;
-  }
-
-  debug(...args: string[]): void {
-    if (this.isDebug) {
-      hilog.debug(this.domain, this.prefix, this.format, args);
-    }
-  }
-
-  info(...args: string[]): void {
-    hilog.info(this.domain, this.prefix, this.format, args);
-  }
-
-  warn(...args: string[]): void {
-    hilog.warn(this.domain, this.prefix, this.format, args);
-  }
-
-  error(...args: string[]): void {
-    hilog.error(this.domain, this.prefix, this.format, args);
-  }
-}
-
-export default new Logger('RNReactNativePingTurboModule', 0xFF00, true)
