--- harmony/ping/src/main/module.json5
@@ -1,22 +1,0 @@
-{
-  "module": {
-    "name": "RNPing",
-    "type": "har",
-    "deviceTypes": [
-      "default",
-      "tablet",
-      "2in1"
-    ],
-    "requestPermissions": [
-      {
-        "name": "ohos.permission.GET_WIFI_INFO"
-      },
-      {
-        "name": "ohos.permission.GET_NETWORK_INFO"
-      },
-      {
-        "name": "ohos.permission.INTERNET"
-      }
-    ]
-  }
-}
