--- harmony/ping/src/main/resources/base/element/string.json
@@ -1,8 +1,0 @@
-{
-  "string": [
-    {
-      "name": "page_show",
-      "value": "page from npm package"
-    }
-  ]
-}
