--- harmony/ping/src/main/resources/zh_CN/element/string.json
@@ -1,8 +1,0 @@
-{
-  "string": [
-    {
-      "name": "page_show",
-      "value": "page from npm package"
-    }
-  ]
-}
